requirements.txt
pandas
numpy
alpaca-trade-api==3.2.0
ta
sqlalchemy
websocket-client
schedule
python-dotenv
loguru
requests
tqdm
matplotlib
plotly
pytest
